Add support for JJB files from multiple repos
To enable JJB definitions to be spread over multiple repositories,
jjb_setup.yml is updated to allow additional repositories to be
specified. The job clones the repositories defined as job sources, the
required paths are determined from the component_metadata.yml files.

The job has been converted to a job template to allow
rpc_jobs/defaults.yml to be used to define the list of job sources in
one place.

JIRA: RE-2069